cpp: move to having a directory per day
I anticipate that having all files in a single directory can be
unwieldy in the long run, especially for more complex problems where
there might be e.g. internal functions that I want to write tests for.
